refactor: improve module coherence in agent decomposition
Move functions to their natural homes for single-responsibility modules:

- Rename _api.py to _cli.py (only CLI/REPL helpers remain)
- Move cancel_run/acancel_run from _tools.py to _run.py
- Move get_agent_data from _tools.py to _storage.py
- Move save_run_response_to_file from _response.py to _storage.py
- Move cleanup_and_store/acleanup_and_store from _cli.py to _storage.py
- Move update_reasoning_content_from_tool_call from _cli.py to _response.py
- Move get_effective_filters from _cli.py to _messages.py
- Move deep_copy/deep_copy_field from _messages.py to _init.py
- Move 12 memory/culture/learning functions from _hooks.py to _storage.py
